improve

+++ main.js
@@ -40,7 +40,7 @@ const db = new SimplDB();
 global.db = db;
 
 // Event handling when the bot is ready.
-bot.ev.once(Events.ClientReady, (m) => {
+bot.ev.once(Events.ClientReady, async (m) => {
     console.log(`Ready at ${m.user.id}`);
     global.system.startTime = Date.now();
 });
@@ -70,23 +70,27 @@ bot.ev.on(Events.MessagesUpsert, async (m, ctx) => {
     // AFK handling: Mentioned users.
     const mentionJids = m.message?.extendedTextMessage?.contextInfo?.mentionedJid;
     if (mentionJids && mentionJids.length > 0) {
-        mentionJids.forEach(async (mentionJid) => {
+        for (const mentionJid of mentionJids) {
             const getAFKMention = db.get(`user.${mentionJid.split("@")[0]}.afk`);
             if (getAFKMention) {
-                const reason = await db.get(`user.${senderNumber}.afk.reason`);
-                const timeStamp = await db.get(`user.${senderNumber}.afk.timeStamp`);
+                const [reason, timeStamp] = await Promise.all([
+                    db.get(`user.${senderNumber}.afk.reason`),
+                    db.get(`user.${senderNumber}.afk.timeStamp`)
+                ]);
                 const timeAgo = smpl.convertMsToDuration(Date.now() - timeStamp);
 
-                ctx.reply(`Dia AFK dengan alasan ${reason} selama ${timeAgo || "kurang dari satu detik."}.`);
+                return ctx.reply(`Dia AFK dengan alasan ${reason} selama ${timeAgo || "kurang dari satu detik."}.`);
             }
-        });
+        }
     }
 
     // AFK handling: Returning from AFK.
     const getAFKMessage = await db.get(`user.${senderNumber}.afk`);
     if (getAFKMessage) {
-        const reason = await db.get(`user.${senderNumber}.afk.reason`);
-        const timeStamp = await db.get(`user.${senderNumber}.afk.timeStamp`);
+        const [reason, timeStamp] = await Promise.all([
+            db.get(`user.${senderNumber}.afk.reason`),
+            db.get(`user.${senderNumber}.afk.timeStamp`)
+        ]);
         const timeAgo = smpl.convertMsToDuration(Date.now() - timeStamp);
         await db.delete(`user.${senderNumber}.afk`);
 
@@ -114,17 +118,7 @@ bot.ev.on(Events.MessagesUpsert, async (m, ctx) => {
             const command = m.content.slice(2);
 
             try {
-                const output = await new Promise((resolve, reject) => {
-                    exec(command, (error, stdout, stderr) => {
-                        if (error) {
-                            reject(new Error(error.message));
-                        } else if (stderr) {
-                            reject(new Error(stderr));
-                        } else {
-                            resolve(stdout);
-                        }
-                    });
-                });
+                const output = await execPromise(command);
 
                 return await ctx.reply(output);
             } catch (error) {
@@ -138,14 +132,11 @@ bot.ev.on(Events.MessagesUpsert, async (m, ctx) => {
     if (isGroup) {
         // Antilink handling.
         const getAntilink = await db.get(`group.${groupNumber}.antilink`);
-        if (getAntilink) {
-            const urlRegex = /((([A-Za-z]{3,9}:(?:\/\/)?)(?:[\-;:&=\+\$,\w]+@)?[A-Za-z0-9\.\-]+|(?:www\.|[\-;:&=\+\$,\w]+@)[A-Za-z0-9\.\-]+)((?:\/[\+~%\/\.\w\-_]*)?\??(?:[\-\+=&;%@\.\w_]*)#?(?:[\.\!\/\\\w]*))?)\b/i;
-            if (m.content && urlRegex.test(m.content)) {
-                if ((await smpl.isAdmin(ctx)) === 1) return;
+        if (getAntilink && m.content && urlRegex.test(m.content)) {
+            if ((await smpl.isAdmin(ctx)) === 1) return;
 
-                await ctx.reply(`${bold("[ ! ]")} Jangan kirim tautan!`);
-                return await ctx.deleteMessage(m.key);
-            }
+            await ctx.reply(`${bold("[ ! ]")} Jangan kirim tautan!`);
+            return await ctx.deleteMessage(m.key);
         }
     }
 
@@ -156,15 +147,7 @@ bot.ev.on(Events.MessagesUpsert, async (m, ctx) => {
         if (getMessageDataMenfess) {
             const from = await db.get(`menfess.${senderNumber}.from`);
             try {
-                await ctx.sendMessage(`${from}@s.whatsapp.net`, {
-                    text: `❖ ${bold("Menfess")}\n` +
-                        `Hai, saya ${global.bot.name}, Dia (${senderNumber}) menjawab pesan menfess yang Anda kirimkan.\n` +
-                        "-----\n" +
-                        `${m.content}\n` +
-                        "-----\n" +
-                        "Jika ingin membalas, Anda harus mengirimkan perintah lagi.\n",
-                });
-                await db.delete(`menfess.${senderNumber}`);
+                await sendMenfess(ctx, senderNumber, from, m.content);
 
                 return ctx.reply("Pesan berhasil terkirim!");
             } catch (error) {
@@ -175,93 +158,82 @@ bot.ev.on(Events.MessagesUpsert, async (m, ctx) => {
     }
 });
 
-// Event handling when a user joins a group.
-bot.ev.on(Events.UserJoin, async (m) => {
-    const {
-        id,
-        participants
-    } = m;
-
-    try {
-        const getWelcome = await db.get(`group.${id.split("@")[0]}.welcome`);
-        if (getWelcome) {
-            const metadata = await bot.core.groupMetadata(id);
-
-            for (const jid of participants) {
-                let profile;
-                try {
-                    profile = await bot.core.profilePictureUrl(jid, "image");
-                } catch {
-                    profile = "https://lh3.googleusercontent.com/proxy/esjjzRYoXlhgNYXqU8Gf_3lu6V-eONTnymkLzdwQ6F6z0MWAqIwIpqgq_lk4caRIZF_0Uqb5U8NWNrJcaeTuCjp7xZlpL48JDx-qzAXSTh00AVVqBoT7MJ0259pik9mnQ1LldFLfHZUGDGY=w1200-h630-p-k-no-nu";
-                }
-
-                await bot.core.sendMessage(id, {
-                    text: `Selamat datang @${jid.split("@")[0]} di grup ${metadata.subject}!`,
-                    contextInfo: {
-                        mentionedJid: [jid],
-                        externalAdReply: {
-                            title: "JOIN",
-                            mediaType: 1,
-                            previewType: 0,
-                            renderLargerThumbnail: true,
-                            thumbnailUrl: profile,
-                            sourceUrl: global.bot.groupChat,
-                        },
-                    },
-                });
-            }
-        }
-    } catch (error) {
-        console.error("Error:", error);
-        return bot.core.sendMessage(id, {
-            text: `${bold("[ ! ]")} Terjadi kesalahan: ${error.message}`,
-        });
-    }
-});
-
-// Event handling when a user leaves a group.
-bot.ev.on(Events.UserLeave, async (m) => {
-    const {
-        id,
-        participants
-    } = m;
-
-    try {
-        const getWelcome = await db.get(`group.${id.split("@")[0]}.welcome`);
-        if (getWelcome) {
-            const metadata = await bot.core.groupMetadata(id);
-
-            for (const jid of participants) {
-                let profile;
-                try {
-                    profile = await bot.core.profilePictureUrl(jid, "image");
-                } catch {
-                    profile = "https://lh3.googleusercontent.com/proxy/esjjzRYoXlhgNYXqU8Gf_3lu6V-eONTnymkLzdwQ6F6z0MWAqIwIpqgq_lk4caRIZF_0Uqb5U8NWNrJcaeTuCjp7xZlpL48JDx-qzAXSTh00AVVqBoT7MJ0259pik9mnQ1LldFLfHZUGDGY=w1200-h630-p-k-no-nu";
-                }
-
-                await bot.core.sendMessage(id, {
-                    text: `@${jid.split("@")[0]} keluar dari grup ${metadata.subject}.`,
-                    contextInfo: {
-                        mentionedJid: [jid],
-                        externalAdReply: {
-                            title: "LEAVE",
-                            mediaType: 1,
-                            previewType: 0,
-                            renderLargerThumbnail: true,
-                            thumbnailUrl: profile,
-                            sourceUrl: global.bot.groupChat,
-                        },
-                    },
-                });
-            }
-        }
-    } catch (error) {
-        console.error("Error:", error);
-        return bot.core.sendMessage(id, {
-            text: `${bold("[ ! ]")} Terjadi kesalahan: ${error.message}`,
-        });
-    }
-});
+// Event handling when a user joins or leaves a group.
+bot.ev.on(Events.UserJoin, handleUserEvent);
+bot.ev.on(Events.UserLeave, handleUserEvent);
 
 // Launch the bot.
 bot.launch().catch((error) => console.error("Error:", error));
+
+// Utility functions
+async function execPromise(command) {
+    return new Promise((resolve, reject) => {
+        exec(command, (error, stdout, stderr) => {
+            if (error) {
+                reject(new Error(error.message));
+            } else if (stderr) {
+                reject(new Error(stderr));
+            } else {
+                resolve(stdout);
+            }
+        });
+    });
+}
+
+async function sendMenfess(ctx, senderNumber, from, content) {
+    await ctx.sendMessage(`${from}@s.whatsapp.net`, {
+        text: `❖ ${bold("Menfess")}\n` +
+            `Hai, saya ${global.bot.name}, Dia (${senderNumber}) menjawab pesan menfess yang Anda kirimkan.\n` +
+            "-----\n" +
+            `${content}\n` +
+            "-----\n" +
+            "Jika ingin membalas, Anda harus mengirimkan perintah lagi.\n",
+    });
+}
+
+async function handleUserEvent(m) {
+    const {
+        id,
+        participants
+    } = m;
+
+    try {
+        const getWelcome = await db.get(`group.${id.split("@")[0]}.welcome`);
+        if (getWelcome) {
+            const metadata = await bot.core.groupMetadata(id);
+
+            for (const jid of participants) {
+                let profile;
+                try {
+                    profile = await bot.core.profilePictureUrl(jid, "image");
+                } catch {
+                    profile = "https://lh3.googleusercontent.com/proxy/esjjzRYoXlhgNYXqU8Gf_3lu6V-eONTnymkLzdwQ6F6z0MWAqIwIpqgq_lk4caRIZF_0Uqb5U8NWNrJcaeTuCjp7xZlpL48JDx-qzAXSTh00AVVqBoT7MJ0259pik9mnQ1LldFLfHZUGDGY=w1200-h630-p-k-no-nu";
+                }
+
+                const message = m.type === Events.UserJoin ? `Selamat datang @${jid.split("@")[0]} di grup ${metadata.subject}!` : `@${jid.split("@")[0]} keluar dari grup ${metadata.subject}.`;
+
+                await bot.core.sendMessage(id, {
+                    text: message,
+                    contextInfo: {
+                        mentionedJid: [jid],
+                        externalAdReply: {
+                            mediaType: 1,
+                            previewType: 0,
+                            mediaUrl: global.bot.groupChat,
+                            title: m.type === Events.UserJoin ? "JOIN" : "LEAVE",
+                            body: null,
+                            renderLargerThumbnail: true,
+                            thumbnailUrl: profile,
+                            sourceUrl: global.bot.groupChat
+                        }
+                    }
+                });
+            }
+        }
+    } catch (error) {
+        console.error("Error:", error);
+        return bot.core.sendMessage(id, {
+            text: `${bold("[ ! ]")} Terjadi kesalahan: ${error.message}`
+        });
+    }
+}
